Removed legacy card and ratings
